Refactor counter redux files to prepare for more reducers
