+++ src/store/actionTypes/counter.ts
